move eval code to different file
