+++ journeys/iteration.py
@@ -28,14 +28,11 @@ from app_utils.shared_utils import (
     get_yamls_from_stage,
     init_session_states,
     return_home_button,
-    schema_selector_container,
     stage_selector_container,
-    table_selector_container,
     upload_yaml,
     validate_and_upload_tmp_yaml,
-    validate_table_exist,
-    validate_table_schema,
 )
+from journeys.evaluation import evaluation_mode_show
 from journeys.joins import joins_dialog
 from semantic_model_generator.data_processing.cte_utils import (
     context_to_column_format,
@@ -48,53 +45,7 @@ from semantic_model_generator.data_processing.proto_utils import (
     yaml_to_semantic_model,
 )
 from semantic_model_generator.protos import semantic_model_pb2
-from semantic_model_generator.snowflake_utils.snowflake_connector import (
-    create_table_in_schema,
-    execute_query,
-    fetch_table,
-    get_table_hash,
-)
 from semantic_model_generator.validate_model import validate
-
-EVALUATION_TABLE_SCHEMA = {
-    "ID": "VARCHAR",
-    "QUERY": "VARCHAR",
-    "GOLD_SQL": "VARCHAR",
-}
-RESULTS_TABLE_SCHEMA = {
-    "TIMESTAMP": "DATETIME",
-    "ID": "VARCHAR",
-    "QUERY": "VARCHAR",
-    "ANALYST_TEXT": "VARCHAR",
-    "ANALYST_SQL": "VARCHAR",
-    "ANALYST_RESULT": "VARCHAR",
-    "GOLD_SQL": "VARCHAR",
-    "GOLD_RESULT": "VARCHAR",
-    "CORRECT": "BOOLEAN",
-    "EXPLANATION": "VARCHAR",
-    "MODEL_HASH": "VARCHAR",
-}
-
-LLM_JUDGE_PROMPT_TEMPLATE = """\
-[INST] Your task is to determine whether the two given dataframes are
-equivalent semantically in the context of a question. You should attempt to
-answer the given question by using the data in each dataframe. If the two
-answers are equivalent, those two dataframes are considered equivalent.
-Otherwise, they are not equivalent. Please also provide your reasoning.
-If they are equivalent, output "REASON: <reason>. ANSWER: true". If they are
-not equivalent, output "REASON: <reason>. ANSWER: false".
-
-### QUESTION: {input_question}
-
-* DATAFRAME 1:
-{frame1_str}
-
-* DATAFRAME 2:
-{frame2_str}
-
-Are the two dataframes equivalent?
-OUTPUT:
-[/INST] """
 
 
 @st.cache_data(show_spinner=False)
@@ -419,151 +370,6 @@ def chat_and_edit_vqr(_conn: SnowflakeConnection) -> None:
         st.session_state.active_suggestion = None
 
 
-def clear_evaluation_data() -> None:
-    session_states = (
-        "eval_table_frame",
-        "eval_table_hash",
-        "selected_eval_database",
-        "selected_eval_schema",
-        "selected_eval_table",
-        "selected_results_eval_database",
-        "selected_results_eval_new_table",
-        "selected_results_eval_new_table_no_schema",
-        "selected_results_eval_old_table",
-        "selected_results_eval_schema",
-        "use_existing_table",
-        "eval_timestamp",
-    )
-    for feature in session_states:
-        if feature in st.session_state:
-            del st.session_state[feature]
-
-
-@st.experimental_dialog("Evaluation Tables", width="large")
-def evaluation_data_dialog() -> None:
-    st.markdown("Please select an evaluation table")
-    table_selector_container(
-        db_selector={"key": "selected_eval_database", "label": "Evaluation database"},
-        schema_selector={"key": "selected_eval_schema", "label": "Evaluation schema"},
-        table_selector={"key": "selected_eval_table", "label": "Evaluation table"},
-    )
-
-    st.divider()
-
-    st.markdown("Please select a results table")
-    eval_results_existing_table = st.checkbox(
-        "Use existing table", key="use_existing_table"
-    )
-
-    if not eval_results_existing_table:
-        schema_selector_container(
-            db_selector={
-                "key": "selected_results_eval_database",
-                "label": "Results database",
-            },
-            schema_selector={
-                "key": "selected_results_eval_schema",
-                "label": "Results schema",
-            },
-        )
-
-        original_new_table_name = st.text_input(
-            key="selected_results_eval_new_table_no_schema",
-            label="Enter the table name to upload evaluation results",
-        )
-        if original_new_table_name:
-            schema_name = st.session_state.get("selected_results_eval_schema")
-            updated_new_table_name = f"{schema_name}.{original_new_table_name}".upper()
-            st.session_state["selected_results_eval_new_table"] = updated_new_table_name
-    else:
-        table_selector_container(
-            db_selector={
-                "key": "selected_results_eval_database",
-                "label": "Results database",
-            },
-            schema_selector={
-                "key": "selected_results_eval_schema",
-                "label": "Results schema",
-            },
-            table_selector={
-                "key": "selected_results_eval_old_table",
-                "label": "Results table",
-            },
-        )
-
-    st.divider()
-
-    if st.button("Use Tables"):
-        st.session_state["selected_results_eval_table"] = st.session_state.get(
-            "selected_results_eval_new_table"
-        ) or st.session_state.get("selected_results_eval_old_table")
-
-        if (
-            not st.session_state["selected_eval_database"]
-            or not st.session_state["selected_eval_schema"]
-            or not st.session_state["selected_eval_table"]
-            or not st.session_state["selected_results_eval_database"]
-            or not st.session_state["selected_results_eval_schema"]
-            or not st.session_state["selected_results_eval_table"]
-        ):
-            st.error("Please fill in all fields.")
-            return
-
-        if not validate_table_schema(
-            table=st.session_state["selected_eval_table"],
-            schema=EVALUATION_TABLE_SCHEMA,
-        ):
-            st.error(f"Evaluation table must have schema {EVALUATION_TABLE_SCHEMA}.")
-            return
-
-        if eval_results_existing_table:
-            if not validate_table_schema(
-                table=st.session_state["selected_results_eval_old_table"],
-                schema=RESULTS_TABLE_SCHEMA,
-            ):
-                st.error(
-                    f"Evaluation result table must have schema {RESULTS_TABLE_SCHEMA}."
-                )
-                return
-
-        else:
-            if validate_table_exist(
-                schema=st.session_state["selected_results_eval_schema"],
-                table_name=st.session_state[
-                    "selected_results_eval_new_table_no_schema"
-                ],
-            ):
-                st.error("Results table already exists")
-                return
-
-            with st.spinner("Creating table..."):
-                success = create_table_in_schema(
-                    conn=get_snowflake_connection(),
-                    table_fqn=st.session_state["selected_results_eval_new_table"],
-                    columns_schema=RESULTS_TABLE_SCHEMA,
-                )
-                if success:
-                    st.success(
-                        f'Table {st.session_state["selected_results_eval_new_table"]} created successfully!'
-                    )
-                else:
-                    st.error(
-                        f'Failed to create table {st.session_state["selected_results_eval_new_table"]}'
-                    )
-                    return
-
-        st.session_state["eval_table_hash"] = get_table_hash(
-            conn=get_snowflake_connection(),
-            table_fqn=st.session_state["selected_eval_table"],
-        )
-        st.session_state["eval_table_frame"] = fetch_table(
-            conn=get_snowflake_connection(),
-            table_fqn=st.session_state["selected_eval_table"],
-        ).set_index("ID")
-
-        st.rerun()
-
-
 @st.experimental_dialog("Upload", width="small")
 def upload_dialog(content: str) -> None:
     def upload_handler(file_name: str) -> None:
@@ -863,385 +669,6 @@ the semantic model must be validated to be uploaded."""
 PARTNER_SEMANTIC_HELP = """Uploaded semantic files from a partner tool?
 Use this feature to integrate partner semantic specs into Cortex Analyst's spec.
 Note that the Cortex Analyst semantic model must be validated before integrating partner semantics."""
-
-
-def evaluation_mode_show() -> None:
-    if st.button("Set Evaluation Tables", on_click=clear_evaluation_data):
-        evaluation_data_dialog()
-
-    if "validated" in st.session_state and not st.session_state["validated"]:
-        st.error("Please validate your semantic model before evaluating.")
-        return
-
-    # TODO: find a less awkward way of specifying this.
-    if any(
-        key not in st.session_state
-        for key in ("selected_eval_table", "eval_table_hash", "eval_table_frame")
-    ):
-        st.error("Please set evaluation tables.")
-        return
-
-    else:
-        results_table = st.session_state.get(
-            "selected_results_eval_old_table"
-        ) or st.session_state.get("selected_results_eval_new_table")
-        st.session_state["eval_timestamp"] = time.strftime("%Y-%m-%d %H:%M:%S")
-        summary_stats = pd.DataFrame(
-            [
-                ["Evaluation Timestamp", st.session_state["eval_timestamp"]],
-                ["Evaluation Table", st.session_state["selected_eval_table"]],
-                ["Evaluation Result Table", results_table],
-                ["Evaluation Table Hash", st.session_state["eval_table_hash"]],
-                ["Semantic Model YAML Hash", hash(st.session_state["working_yml"])],
-                ["Query Count", len(st.session_state["eval_table_frame"])],
-            ],
-            columns=["Summary Statistic", "Value"],
-        )
-        st.dataframe(summary_stats, hide_index=True)
-
-        send_analyst_requests()
-        run_sql_queries()
-        result_comparisons()
-
-
-def send_analyst_requests() -> None:
-    def _get_content(
-        x: dict, item_type: str, key: str, default: str = ""  # type: ignore[type-arg]
-    ) -> str:
-        result = next(
-            (
-                item[key]
-                for item in x["message"]["content"]
-                if item["type"] == item_type
-            ),
-            default,
-        )
-        return result
-
-    eval_table_frame: pd.DataFrame = st.session_state["eval_table_frame"]
-
-    total_requests = len(eval_table_frame)
-    progress_bar = st.progress(0)
-    status_text = st.empty()
-    start_time = time.time()
-    analyst_results = []
-
-    for i, (row_id, row_id) in enumerate(eval_table_frame.iterrows(), start=1):
-        status_text.text(f"Sending request {i}/{total_requests} to Analyst...")
-        messages = [
-            {"role": "user", "content": [{"type": "text", "text": row_id["QUERY"]}]}
-        ]
-        semantic_model = proto_to_yaml(st.session_state.semantic_model)
-        try:
-            response = send_message(
-                _conn=get_snowflake_connection(),
-                semantic_model=semantic_model,
-                messages=messages,  # type: ignore[arg-type]
-            )
-            response_text = _get_content(response, item_type="text", key="text")
-            response_sql = _get_content(response, item_type="sql", key="statement")
-            analyst_results.append(
-                dict(ID=row_id, ANALYST_TEXT=response_text, ANALYST_SQL=response_sql)
-            )
-        except Exception as e:
-            import traceback
-
-            st.error(f"Problem with {row_id}: {e} \n{traceback.format_exc()}")
-
-        progress_bar.progress(i / total_requests)
-        time.sleep(0.1)
-
-    elapsed_time = time.time() - start_time
-    status_text.text(
-        f"All analyst requests received ✅ (Time taken: {elapsed_time:.2f} seconds)"
-    )
-
-    analyst_results_frame = pd.DataFrame(analyst_results).set_index("ID")
-    st.session_state["analyst_results_frame"] = analyst_results_frame
-
-
-def run_sql_queries() -> None:
-    eval_table_frame: pd.DataFrame = st.session_state["eval_table_frame"]
-    analyst_results_frame = st.session_state["analyst_results_frame"]
-
-    total_requests = len(eval_table_frame)
-    progress_bar = st.progress(0)
-    status_text = st.empty()
-    start_time = time.time()
-
-    analyst_results = []
-    gold_results = []
-
-    for i, (row_id, eval_row) in enumerate(eval_table_frame.iterrows(), start=1):
-        status_text.text(f"Evaluating Analyst query {i}/{total_requests}...")
-
-        analyst_query = analyst_results_frame.loc[row_id, "ANALYST_SQL"]
-        analyst_result = execute_query(
-            conn=get_snowflake_connection(), query=analyst_query
-        )
-        analyst_results.append(analyst_result)
-
-        gold_query = eval_table_frame.loc[row_id, "GOLD_SQL"]
-        gold_result = execute_query(conn=get_snowflake_connection(), query=gold_query)
-        gold_results.append(gold_result)
-
-        progress_bar.progress(i / total_requests)
-        time.sleep(0.1)
-
-    st.session_state["query_results_frame"] = pd.DataFrame(
-        data=dict(ANALYST_RESULT=analyst_results, GOLD_RESULT=gold_results),
-        index=eval_table_frame.index,
-    )
-
-    elapsed_time = time.time() - start_time
-    status_text.text(
-        f"All analyst and gold queries run ✅ (Time taken: {elapsed_time:.2f} seconds)"
-    )
-
-
-def _match_series(analyst_frame: pd.DataFrame, gold_series: pd.Series) -> str | None:
-    """Determine which result frame column name matches the gold series.
-
-    Args:
-        analyst_frame: the data generated from the LLM constructed user query
-        gold_series: a column from the data generated from the gold sql
-
-    Returns:
-        if there is a match, the results column name, if not, None
-    """
-    for analyst_col in analyst_frame:
-        assert isinstance(analyst_col, str)
-        try:
-            pd.testing.assert_series_equal(
-                left=analyst_frame[analyst_col],
-                right=gold_series,
-                check_names=False,
-            )
-            return analyst_col
-        except AssertionError:
-            pass
-
-    return None
-
-
-def _results_contain_gold_data(
-    analyst_frame: pd.DataFrame,
-    gold_frame: pd.DataFrame,
-) -> bool:
-    """Determine if result frame contains all the same values as a gold frame.
-
-    Args:
-        analyst_frame: the data generated from the LLM constructed user query
-        gold_frame: the data generated from a gold sql query
-
-    Returns:
-        a boolean indicating if the results contain the gold data
-    """
-    if analyst_frame.shape[0] != gold_frame.shape[0]:
-        return False
-
-    unmatched_result_cols = analyst_frame.columns
-    for gold_col in gold_frame:
-        matching_col = _match_series(
-            analyst_frame=analyst_frame[unmatched_result_cols],
-            gold_series=gold_frame[gold_col],
-        )
-        if matching_col is None:
-            return False
-        else:
-            unmatched_result_cols = unmatched_result_cols.drop(matching_col)
-
-    return True
-
-
-def _llm_judge(frame: pd.DataFrame) -> pd.DataFrame:
-
-    if frame.empty:
-        return pd.DataFrame({"EXPLANATION": [], "CORRECT": []})
-
-    # create prompt frame series
-    table_name = "__LLM_JUDGE_TEMP_TABLE"
-    col_name = "LLM_JUDGE_PROMPT"
-
-    prompt_frame = frame.apply(
-        axis=1,
-        func=lambda x: LLM_JUDGE_PROMPT_TEMPLATE.format(
-            input_question=x["QUERY"],
-            frame1_str=x["ANALYST_RESULT"].to_string(index=False),
-            frame2_str=x["GOLD_RESULT"].to_string(index=False),
-        ),
-    ).to_frame(name=col_name)
-    conn = get_snowflake_connection()
-    _ = write_pandas(
-        conn=conn,
-        df=prompt_frame,
-        table_name=table_name,
-        auto_create_table=True,
-        table_type="temporary",
-        overwrite=True,
-    )
-
-    query = f"""
-    SELECT SNOWFLAKE.CORTEX.COMPLETE('mistral-large2', {col_name}) AS LLM_JUDGE
-    FROM {conn.database}.{conn.schema}.{table_name}
-    """
-    cursor = conn.cursor()
-    cursor.execute(query)
-    llm_judge_frame = cursor.fetch_pandas_all()
-    llm_judge_frame.index = frame.index
-
-    reason_filter = re.compile(r"REASON\:([\S\s]*?)ANSWER\:")
-    answer_filter = re.compile(r"ANSWER\:([\S\s]*?)$")
-
-    def _safe_re_search(x, filter):  # type: ignore[no-untyped-def]
-        try:
-            return re.search(filter, x).group(1).strip()  # type: ignore[union-attr]
-        except Exception as e:
-            return f"Could Not Parse LLM Judge Response: {x} with error: {e}"
-
-    llm_judge_frame["EXPLANATION"] = llm_judge_frame["LLM_JUDGE"].apply(
-        _safe_re_search, args=(reason_filter,)
-    )
-    llm_judge_frame["CORRECT"] = (
-        llm_judge_frame["LLM_JUDGE"]
-        .apply(_safe_re_search, args=(answer_filter,))
-        .str.lower()
-        .eq("true")
-    )
-    return llm_judge_frame
-
-
-def visualize_eval_results(frame: pd.DataFrame) -> None:
-    n_questions = len(frame)
-    n_correct = frame["CORRECT"].sum()
-    accuracy = (n_correct / n_questions) * 100
-    st.markdown(
-        f"###### Results: {n_correct} out of {n_questions} questions correct with accuracy {accuracy:.2f}%"
-    )
-    for id, frame_row in frame.iterrows():
-        match_emoji = "✅" if row["CORRECT"] else "❌"
-        with st.expander(f"Row ID: {id} {match_emoji}"):
-            st.write(f"Input Query: {frame_row['QUERY']}")
-            st.write(frame_row["ANALYST_TEXT"].replace("\n", " "))
-
-            col1, col2 = st.columns(2)
-
-            with col1:
-                st.write("Analyst SQL")
-                st.code(frame_row["ANALYST_SQL"], language="sql")
-
-            with col2:
-                st.write("Golden SQL")
-                st.code(frame_row["GOLD_SQL"], language="sql")
-
-            col1, col2 = st.columns(2)
-            with col1:
-                if isinstance(frame_row["ANALYST_RESULT"], str):
-                    st.error(frame_row["ANALYST_RESULT"])
-                else:
-                    st.write(frame_row["ANALYST_RESULT"])
-
-            with col2:
-                if isinstance(frame_row["GOLD_RESULT"], str):
-                    st.error(frame_row["GOLD_RESULT"])
-                else:
-                    st.write(frame_row["GOLD_RESULT"])
-
-            st.write(f"**Explanation**: {frame_row['EXPLANATION']}")
-
-
-def result_comparisons() -> None:
-    eval_table_frame: pd.DataFrame = st.session_state["eval_table_frame"]
-    analyst_results_frame = st.session_state["analyst_results_frame"]
-    query_results_frame = st.session_state["query_results_frame"]
-
-    frame = pd.concat(
-        [eval_table_frame, analyst_results_frame, query_results_frame], axis=1
-    )
-    start_time = time.time()
-    status_text = st.empty()
-
-    matches = pd.Series(False, index=frame.index)
-    explanations = pd.Series("", index=frame.index)
-    use_llm_judge = "<use llm judge>"
-
-    status_text.text("Checking for exact matches...")
-    for row_id, res_row in frame.iterrows():
-        analyst_is_frame = isinstance(res_row["ANALYST_RESULT"], pd.DataFrame)
-        gold_is_frame = isinstance(res_row["GOLD_RESULT"], pd.DataFrame)
-        if (not analyst_is_frame) and (not gold_is_frame):
-            matches[row_id] = False
-            explanations[row_id] = dedent(
-                f"""
-                analyst sql had an error: {res_row["ANALYST_RESULT"]}
-                gold sql had an error: {res_row["GOLD_RESULT"]}
-                """
-            )
-        elif (not analyst_is_frame) and gold_is_frame:
-            matches[row_id] = False
-            explanations[row_id] = dedent(
-                f"""
-                analyst sql had an error: {res_row["ANALYST_RESULT"]}
-                """
-            )
-        elif analyst_is_frame and (not gold_is_frame):
-            matches[row_id] = False
-            explanations[row_id] = dedent(
-                f"""
-                gold sql had an error: {res_row["GOLD_RESULT"]}
-                """
-            )
-        else:
-            exact_match = _results_contain_gold_data(
-                analyst_frame=res_row["ANALYST_RESULT"],
-                gold_frame=res_row["GOLD_RESULT"],
-            )
-            matches[row_id] = exact_match
-            explanations[row_id] = (
-                "Data matches exactly" if exact_match else use_llm_judge
-            )
-
-    frame["CORRECT"] = matches
-    frame["EXPLANATION"] = explanations
-
-    filtered_frame = frame[explanations == use_llm_judge]
-
-    status_text.text("Calling LLM Judge...")
-    llm_judge_frame = _llm_judge(frame=filtered_frame)
-
-    for col in ("CORRECT", "EXPLANATION"):
-        frame[col] = llm_judge_frame[col].combine_first(frame[col])
-
-    elapsed_time = time.time() - start_time
-    status_text.text(
-        f"Analyst and Gold Results Compared ✅ (Time taken: {elapsed_time:.2f} seconds)"
-    )
-
-    visualize_eval_results(frame)
-
-    frame["TIMESTAMP"] = st.session_state["eval_timestamp"]
-    frame["EVAL_TABLE"] = st.session_state["selected_eval_table"]
-    frame["EVAL_TABLE_HASH"] = st.session_state["eval_table_hash"]
-    frame["MODEL_HASH"] = hash(st.session_state["working_yml"])
-
-    # Save results to frame as string
-    frame["ANALYST_RESULT"] = frame["ANALYST_RESULT"].apply(
-        lambda x: x.to_string(index=False) if isinstance(x, pd.DataFrame) else x
-    )
-    frame["GOLD_RESULT"] = frame["GOLD_RESULT"].apply(
-        lambda x: x.to_string(index=False) if isinstance(x, pd.DataFrame) else x
-    )
-
-    frame = frame.reset_index()[list(RESULTS_TABLE_SCHEMA)]
-    write_pandas(
-        conn=get_snowflake_connection(),
-        df=frame,
-        table_name=st.session_state["selected_results_eval_table"],
-        overwrite=False,
-        quote_identifiers=False,
-        auto_create_table=False,
-    )
-    st.write("Evaluation results stored in the database ✅")
 
 
 def show() -> None:
